add few more flags to enrich tenants selectively
